move aiqpd testing to test_services instead of test_core
