+++ dodola/tests/test_core.py
@@ -1,17 +1,12 @@
 import numpy as np
-from numpy.testing import assert_approx_equal
 import pytest
-import pandas as pd
 import xarray as xr
 import cftime
-from xclim.core.calendar import convert_calendar
-from xclim.sdba.adjustment import QuantileDeltaMapping
 from xclim.sdba.utils import equally_spaced_nodes
 from xclim import sdba, set_options
 from dodola.core import (
     train_quantiledeltamapping,
     adjust_quantiledeltamapping_year,
-    train_analogdownscaling,
 )
 
 
@@ -187,78 +182,4 @@ def test_adjust_quantiledeltamapping_year_output_time():
     )
     assert max(adjusted_ds["time"].values) == cftime.DatetimeNoLeap(
         2088, 12, 31, 0, 0, 0, 0
-    )
-
-
-def test_analoginspired_quantilepreserving_downscaling():
-    """Tests that the average of AIQPD values equals the bias corrected
-    value for the corresponding coarse-res gridcells"""
-    # make test data
-    np.random.seed(0)
-    lon = [-99.83, -99.32, -99.79, -99.23]
-    lat = [42.25, 42.21, 42.63, 42.59]
-    # TO-DO: update time range to include +/- 15 days
-    time = pd.date_range(start="1995-01-01", end="2014-12-31")
-    temperature = 15 + 8 * np.random.randn(len(time), 4, 4)
-
-    ds = xr.Dataset(
-        data_vars=dict(
-            air=(["time", "lat", "lon"], temperature),
-        ),
-        coords=dict(
-            time=time,
-            lon=(["lon"], lon),
-            lat=(["lat"], lat),
-        ),
-        attrs=dict(description="Weather related data."),
-    )
-
-    # remove leap days and only use four gridcells
-    temp_slice = convert_calendar(ds["air"], target="noleap")
-
-    # take the mean across space to represent coarse reference data for AFs
-    temp_slice_mean = temp_slice.mean(["lat", "lon"])
-    # then tile it to be on the same grid as the fine reference data
-    temp_slice_mean_resampled = temp_slice_mean.broadcast_like(temp_slice)
-
-    # need to create some fake bias corrected data so that we can use it to downscale
-    with set_options(sdba_extra_output=True):
-        quantiles = equally_spaced_nodes(620, eps=None)
-        QDM = QuantileDeltaMapping(
-            kind="+",
-            nquantiles=quantiles,
-            group=sdba.Grouper("time.dayofyear", window=31),
-        )
-        QDM.train(temp_slice_mean + 2, temp_slice_mean)
-        fake_biascorrected = QDM.adjust(temp_slice_mean + 4)
-        # this is necessary to make sim_q a coordinate on 'scen'
-        fake_biascorrected = (
-            fake_biascorrected["scen"]
-            .assign_coords(sim_q=fake_biascorrected.sim_q)
-            .to_dataset()
-        )
-
-    # now downscale
-    # note that we are testing with 62 quantiles because we
-    # only have two years of data
-    aiqpd = train_analogdownscaling(
-        temp_slice_mean_resampled.to_dataset(name="scen"),
-        temp_slice.to_dataset(name="scen"),
-        variable="scen",
-        kind="+",
-        quantiles_n=620,
-    )
-
-    # make bias corrected data on the fine resolution grid
-    biascorrected = fake_biascorrected["scen"].broadcast_like(temp_slice)
-    # downscale the bias corrected data
-    aiqpd_downscaled = aiqpd.adjust(biascorrected)
-
-    # check that bias corrected value at a given timestep equals the average
-    # of the downscaled values that correspond to the bias corrected value
-    bias_corrected_value = biascorrected.isel(time=100).values[0][0]
-    downscaled_average = aiqpd_downscaled.isel(time=100).mean().values
-
-    assert_approx_equal(
-        bias_corrected_value, downscaled_average, significant=5, verbose=True
     )
